requirements.txt
Appium-Python-Client==2.7.1
asserts==0.12.0
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
ddt==1.6.0
decorator==5.1.1
docutils==0.19
exceptiongroup==1.0.4
future==0.18.2
h11==0.14.0
idna==3.4
Jinja2==3.1.2
jsonpath-ng==1.5.3
kitchen==1.2.6
MarkupSafe==2.1.1
mysql-connector-python==8.0.32
natsort==8.2.0
oauthlib==3.2.2
outcome==1.2.0
pefile==2022.5.30
pluggy==1.0.0
ply==3.11
protobuf==3.20.3
pycparser==2.21
pydantic==1.10.2
Pygments==2.14.0
PySocks==1.7.1
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
robotframework==5.0.1
robotframework-appiumlibrary==2.0.0
robotframework-datadriver==1.7.0
robotframework-extendedrequestslibrary==0.5.5
robotframework-pabot==2.8.0
robotframework-pythonlibcore==4.0.0
robotframework-requests==0.9.4
robotframework-seleniumlibrary==6.0.0
robotframework-stacktrace==0.4.1
selenium==4.6.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.13
wsproto==1.2.0
